refactor: extract bookmark UI logic from MainController
